(svn r22522) -Fix (r22489): (size_t)(uint32)(-1) != (size_t)-1

+++ src/settings_internal.h
@@ -56,7 +56,7 @@ typedef SimpleTinyEnumT<SettingGuiFlagLong, uint16> SettingGuiFlag;
 
 
 typedef bool OnChange(int32 var);           ///< callback prototype on data modification
-typedef int32 OnConvert(const char *value); ///< callback prototype for convertion error
+typedef size_t OnConvert(const char *value); ///< callback prototype for convertion error
 
 /** Properties of config file settings. */
 struct SettingDescBase {
